migrate team controller to new controller architecture
