+++ api/src/controllers/index.ts
@@ -8,5 +8,5 @@
 export * from "./conversation.controller/";
 export * from "./inbound.controller/";
 export * from "./integration.controller/";
-export * from "./team.controller";
+export * from "./team.controller/";
 export * from "./user.controller";
